% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle_theme.R
\name{toggle_theme}
\alias{toggle_theme}
\title{Toggle RStudio themes}
\usage{
toggle_theme(themes = getOption("themetoggle.themes"))
}
\arguments{
\item{themes}{a vector of theme names to toggle through}
}
\description{
Applies an RStudio them on invocation, cycling through given themes.
}
